requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
openai==0.27.4
chromadb